feat: SPLAT-582: Adjust configs.
